bentoml-service/service.py
import bentoml
from bentoml.io import JSON, Image, Text, File
from PIL import Image as PILImage
import io
import base64
# In a real scenario, you would import your model runners here
# from my_models import moondream_runner, melotts_runner

# Mock runners for demonstration purposes if actual models aren't present in this environment
class MoondreamRunnable(bentoml.Runnable):
    SUPPORTED_RESOURCES = ("cpu", "nvidia.com/gpu")
    SUPPORTS_CPU_MULTI_THREADING = True

    def __init__(self):
        # Initialize model here
        pass

    @bentoml.Runnable.method(batchable=False)
    def analyze(self, image: PILImage.Image):
        # Placeholder logic
        return {"description": "A person holding a white cane standing on a sidewalk.", "confidence": 0.95}

class MeloTTSRunnable(bentoml.Runnable):
    SUPPORTED_RESOURCES = ("cpu", "nvidia.com/gpu")
    SUPPORTS_CPU_MULTI_THREADING = True

    def __init__(self):
        # Initialize model here
        pass

    @bentoml.Runnable.method(batchable=False)
    def synthesize(self, text: str):
        # Placeholder: return dummy audio bytes
        # In reality, this returns wav/mp3 bytes
        return b"RIFF....WAVEfmt ...." 

moondream_runner = bentoml.Runner(MoondreamRunnable, name="moondream_runner")
melotts_runner = bentoml.Runner(MeloTTSRunnable, name="melotts_runner")

svc = bentoml.Service("be_my_eyes_ai", runners=[moondream_runner, melotts_runner])

@svc.api(input=Image(), output=JSON())
def vision_analyze(image: PILImage.Image):
    result = moondream_runner.analyze.run(image)
    return result

@svc.api(input=Text(), output=File(mime_type="audio/wav"))
def audio_tts(text: str):
    audio_data = melotts_runner.synthesize.run(text)
    return io.BytesIO(audio_data)

@svc.api(input=Text(), output=JSON())
def status(input_data: str):
    return {"status": "healthy", "models": ["moondream", "melotts"]}

