bentoml-service/requirements.txt
bentoml
torch
torchvision
transformers
pillow
numpy

